Organize CMakeLists and structure in namespaces
