--- Map/MapDriver.cpp
@@ -1,10 +1,0 @@
-#include <iostream>
-#include "Map.h"
-
-using namespace std;
-
-int run(){
-    cout << "Map Unit test" << endl;
-
-    return 0;
-};
