+++ src/Map/CMakeLists.txt
@@ -1,0 +1,4 @@
+list(APPEND SRC_FILES
+        ${CMAKE_CURRENT_LIST_DIR}/Map.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/MapDriver.cpp
+        )
