+++ src/Map/MapDriver.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace Map {
+    namespace Driver {
+        bool run();
+    }
+}
