+++ src/main.cpp
@@ -1,9 +1,11 @@
 #include <iostream>
+#include "Map/MapDriver.h"
 using namespace std;
 
 int main() {
-
     cout << "Domination Game - Team 12 - COMP 345" << endl;
+
+    Map::Driver::run();
 
     return 0;
 }
